MNT: refactor walk helpers into a separate submodule
